wip testing crossing point logic
