+++ src/modules/createSvg/createSvgTypes.ts
@@ -1,0 +1,4 @@
+export type ProcessedPoint = {
+  lat: number;
+  long: number;
+};
